refactor(span): remove `source_type` module (#11024)
It's unnecessary to have a separate `errors.rs` file for `SourceType`. Combine 2 files into 1.

Pure refactor. This PR doesn't alter any of the logic, just relocates it.
