+++ tasks/ast_tools/src/main.rs
@@ -227,7 +227,7 @@ static SOURCE_PATHS: &[&str] = &[
     "crates/oxc_syntax/src/symbol.rs",
     "crates/oxc_syntax/src/reference.rs",
     "crates/oxc_span/src/span.rs",
-    "crates/oxc_span/src/source_type/mod.rs",
+    "crates/oxc_span/src/source_type.rs",
     "crates/oxc_regular_expression/src/ast.rs",
     "napi/parser/src/raw_transfer_types.rs",
 ];
